Migrate base model class testing to pytest
