+++ tests/CMakeLists.txt
@@ -44,7 +44,6 @@ if(RUN_CORE_TESTS)
   add_python_test(item)
   add_python_test(logging_config)
   add_python_test(mail)
-  add_python_test(model)
   add_python_test(model_singleton)
   add_python_test(routetable)
   add_python_test(notification TIMEOUT 50)
